Add try, except block to create_db
